chore(server): move ContactService interface from app/ to identity/
`ContactService` was an interface defined in `app/app-host.ts` (where
it was used as a registered slot) but implemented in `adapters/webhook-
contact-service.ts`. That created an adapters→app reverse-layer
dependency that violates the layered architecture.

The interface is a contact-policy contract — domain-wise it belongs in
identity/. Moving it there:

- Adapters now imports from identity/ (correct layering: adapters can
  reach into identity, the lower layer).
- App imports the interface from identity/ for its registered-slot type
  (app's allowed to read from any lower layer).
- The interface is distinct from `ContactsService` (the CRUD class in
  `identity/services/contact.service.ts`); a new file
  `identity/services/contact-policy.ts` houses it with a comment
  explaining the distinction.

Changes:
- New file: `identity/services/contact-policy.ts` (17 lines).
- `app/app-host.ts`: drop the interface; import from identity/.
- `app/types.ts`: update import path.
- `adapters/webhook-contact-service.ts`: update import path.

Tests: 22/22 pass in src/app.
